# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/logs": {
      "get": {
        "operationId": "LoggerController_getLogs",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "조회할 페이지 번호",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "itemsPerPage",
            "required": true,
            "in": "query",
            "description": "페이지당 조회할 항목 수",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "levels",
            "required": false,
            "in": "query",
            "description": "필터링할 로그 레벨 목록",
            "schema": {
              "minItems": 1,
              "maxItems": 10,
              "example": [
                "error",
                "warn"
              ],
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "조회 시작 일자 (YYYY-MM-DD 또는 YYYY/MM/DD)",
            "schema": {
              "example": "2024-01-01",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "조회 종료 일자 (YYYY-MM-DD 또는 YYYY/MM/DD)",
            "schema": {
              "example": "2024-01-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "로그 목록을 조회합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLogsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "logs"
        ]
      },
      "delete": {
        "operationId": "LoggerController_deleteLogs",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteLogsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "선택한 로그 또는 조건에 맞는 로그를 삭제합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteLogsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "logs"
        ]
      }
    },
    "/logs/{id}": {
      "get": {
        "operationId": "LoggerController_getLogDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "조회 대상 로그 ID",
            "schema": {
              "minimum": 1,
              "example": 42,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "특정 로그의 상세 정보를 조회합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetLogDetailResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "logs"
        ]
      }
    },
...
    "/settings/{key}": {
      "get": {
        "operationId": "SettingsController_getSetting",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "description": "설정을 조회하거나 수정할 때 사용하는 키",
            "schema": {
              "example": "uiLanguage",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "설정 값을 조회합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSettingResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "SettingsController_updateSetting",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "description": "설정을 조회하거나 수정할 때 사용하는 키",
            "schema": {
              "example": "uiLanguage",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "설정 값을 생성하거나 업데이트합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSettingResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "settings"
        ]
      },
      "delete": {
        "operationId": "SettingsController_deleteSetting",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "description": "설정을 조회하거나 수정할 때 사용하는 키",
            "schema": {
              "example": "uiLanguage",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "설정을 삭제합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteSettingResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "settings"
        ]
      }
    },
...
  },
  "info": {
    "title": "FormattedDocs API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "LogListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "로그 ID",
            "example": 1
          },
          "level": {
            "type": "string",
            "description": "로그 레벨",
            "example": "error"
          },
          "message": {
            "type": "string",
            "description": "로그 메시지",
            "example": "Unhandled exception occurred."
          },
          "context": {
            "type": "string",
            "description": "로그 컨텍스트",
            "nullable": true,
            "example": "ImageTranslatorService"
          },
          "metadataPreview": {
            "type": "string",
            "description": "메타데이터 미리보기 문자열",
            "nullable": true,
            "example": "{\"file\":\"example.png\"}"
          },
          "hasMetadata": {
            "type": "boolean",
            "description": "메타데이터 존재 여부",
            "example": true
          },
          "timestamp": {
            "type": "string",
            "description": "로그 발생 시각 (ISO 문자열)",
            "example": "2024-01-01T12:34:56.000Z"
          }
        },
        "required": [
          "id",
          "level",
          "message",
          "hasMetadata",
          "timestamp"
        ]
      },
      "GetLogsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          },
          "logs": {
            "description": "로그 목록",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LogListItemDto"
            }
          },
          "totalItems": {
            "type": "number",
            "description": "총 로그 개수",
            "example": 120
          }
        },
        "required": [
          "success",
          "logs",
          "totalItems"
        ]
      },
      "LogDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "로그 ID",
            "example": 1
          },
          "level": {
            "type": "string",
            "description": "로그 레벨",
            "example": "error"
          },
          "message": {
            "type": "string",
            "description": "로그 메시지",
            "example": "Unhandled exception occurred."
          },
          "context": {
            "type": "string",
            "description": "로그 컨텍스트",
            "nullable": true,
            "example": "ImageTranslatorService"
          },
          "metadataPreview": {
            "type": "string",
            "description": "메타데이터 미리보기 문자열",
            "nullable": true,
            "example": "{\"file\":\"example.png\"}"
          },
          "hasMetadata": {
            "type": "boolean",
            "description": "메타데이터 존재 여부",
            "example": true
          },
          "timestamp": {
            "type": "string",
            "description": "로그 발생 시각 (ISO 문자열)",
            "example": "2024-01-01T12:34:56.000Z"
          },
          "metadata": {
            "type": "string",
            "description": "로그 전체 메타데이터(JSON 문자열)",
            "nullable": true,
            "example": "{\"file\":\"example.png\",\"line\":42}"
          },
          "stack": {
            "type": "string",
            "description": "스택 트레이스 문자열",
            "nullable": true,
            "example": "Error: Something bad\n    at ..."
          },
          "meta": {
            "type": "object",
            "description": "파싱된 메타데이터 객체",
            "nullable": true,
            "example": {
              "file": "example.png",
              "line": 42
            }
          }
        },
        "required": [
          "id",
          "level",
          "message",
          "hasMetadata",
          "timestamp"
        ]
      },
      "GetLogDetailResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          },
          "log": {
            "description": "요청한 로그 상세 정보",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/LogDetailDto"
              }
            ]
          }
        },
        "required": [
          "success"
        ]
      },
      "LogSearchParamsDto": {
        "type": "object",
        "properties": {
          "levels": {
            "description": "필터링할 로그 레벨 목록",
            "example": [
              "error",
              "warn"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "startDate": {
            "type": "string",
            "description": "조회 시작 일자 (YYYY-MM-DD 또는 YYYY/MM/DD)",
            "example": "2024-01-01"
          },
          "endDate": {
            "type": "string",
            "description": "조회 종료 일자 (YYYY-MM-DD 또는 YYYY/MM/DD)",
            "example": "2024-01-31"
          }
        }
      },
      "DeleteLogsRequestDto": {
        "type": "object",
        "properties": {
          "logIds": {
            "description": "삭제할 로그 ID 목록",
            "example": [
              1,
              2,
              3
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "searchParams": {
            "description": "검색 조건과 일치하는 로그를 삭제할 때 사용하는 필터",
            "allOf": [
              {
                "$ref": "#/components/schemas/LogSearchParamsDto"
              }
            ]
          }
        }
      },
      "DeleteLogsResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          }
        },
        "required": [
          "success"
        ]
      },
...
      "GetSettingResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          },
          "result": {
            "type": "string",
            "description": "해당 키에 저장된 설정 값",
            "example": "ko",
            "nullable": true
          }
        },
        "required": [
          "success"
        ]
      },
      "UpdateSettingRequestDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "description": "저장할 설정 값",
            "example": "ko"
          }
        },
        "required": [
          "value"
        ]
      },
      "AppSettingDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "description": "설정 키",
            "example": "uiLanguage"
          },
          "value": {
            "type": "string",
            "description": "설정 값",
            "example": "ko"
          },
          "createdAt": {
            "type": "string",
            "description": "설정이 생성된 시각",
            "example": "2024-04-01T12:34:56.000Z",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "설정이 마지막으로 업데이트된 시각",
            "example": "2024-04-03T08:15:30.000Z",
            "format": "date-time"
          }
        },
        "required": [
          "key",
          "value",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateSettingResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          },
          "result": {
            "description": "저장된 설정 정보",
            "allOf": [
              {
                "$ref": "#/components/schemas/AppSettingDto"
              }
            ]
          }
        },
        "required": [
          "success",
          "result"
        ]
      },
...
      "DeleteSettingResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청이 성공했는지 여부",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "요청 처리 결과 메시지",
            "example": "처리가 완료되었습니다."
          }
        },
        "required": [
          "success"
        ]
      },
...
    }
  }
}